ultibo-retro-run.sh checks for successful kernel update

+++ ultibo-retro-run.sh
@@ -1,3 +1,8 @@
-sudo mv /boot/kernel7.img /boot/kernel7_l.img
-sudo cp /home/pi/ultibo-retro/kernel7.img /boot/kernel7.img
-sudo reboot
+sudo cp /boot/kernel7.img /boot/kernel7_l.img
+sudo cp kernel7.img /boot/kernel7.img
+if [[ "$?" == "0" ]]
+then
+    sudo reboot
+else
+    echo unsuccessful update of /boot/kernel7.img
+fi
